requirements.txt
awscli==1.22.49
boto3==1.20.46
robotframework== 4.1.3
python-dotenv==22.0.3